Update local setup script and update Go dependencies
This commit includes changes to the local setup script to use blocks of concrete account names and hardcoded accounts instead of using an array of keys. It also adds a new 'sequencer' start command and updates several Go dependencies. Removed unnecessary lines of text in README.md and junctionUtils.go.